added some styling
