+++ src/components/Activities/activitiesLayout.js
@@ -6,9 +6,9 @@ import injectTapEventPlugin from "react-tap-event-plugin";
 const muiTheme = getMuiTheme();
 
 
-const styles = require('./styles').popStyle
-const customContentStyle = require('./styles').customContentStyle
-const backgroundStyle = require('./styles').bodyStyle
+const styles = require('../styles').popStyle
+const customContentStyle = require('../styles').customContentStyle
+const backgroundStyle = require('../styles').bodyStyle
 
 class Activities extends React.Component {
   constructor(props){
