+++ src/components/Header/header.css
@@ -1,0 +1,23 @@
+@import url('https://fonts.googleapis.com/css?family=Raleway');
+.header{
+    top:0;
+    margin:0;
+    background-color: #162238;
+    color:#9AA1A7;
+    height:70px;
+    width:100%;
+    display:flex;
+    justify-content: space-between
+}
+.nav{
+    display: flex;
+    width: 400px;
+    justify-content: space-around;
+    margin-right:35px;
+    margin-top:15px;
+    /* font-size:15px; */
+}
+.logo{
+    font-family: 'Raleway', sans-serif;
+    font-size: 50px;
+}
