+++ src/components/Job/Jobs.js
@@ -2,22 +2,22 @@ import React, { Component } from 'react';
 import Layout from "./jobLayout";
 import injectTapEventPlugin from 'react-tap-event-plugin';
 
-const Info = ({ dataArr }) => {
-    return (
-        <div>
-            {dataArr.map((el) => {
-                return <p> {`Agency: ${el.agency}`} <br></ br>
-                           {`Address: ${el.address}`} <br></ br>
-                           {`Borough community: ${el.borough_community}`} <br></ br>
-                           {`Contact Number: ${el.contact_number}`} <br></ br>
-                           {`Program: ${el.program}`} <br></ br>
-                           {`Program Type: ${el.program_type}`} <br></ br>
-                           {`Site name: ${el.site_name}`} <br></ br>
-                        </p>
-                })}
-        </div>
-    )
-};
+// const Info = ({ dataArr }) => {
+//     return (
+//         <div>
+//             {dataArr.map((el) => {
+//                 return <p> {`Agency: ${el.agency}`} <br></ br>
+//                            {`Address: ${el.address}`} <br></ br>
+//                            {`Borough community: ${el.borough_community}`} <br></ br>
+//                            {`Contact Number: ${el.contact_number}`} <br></ br>
+//                            {`Program: ${el.program}`} <br></ br>
+//                            {`Program Type: ${el.program_type}`} <br></ br>
+//                            {`Site name: ${el.site_name}`} <br></ br>
+//                         </p>
+//                 })}
+//         </div>
+//     )
+// };
 
 
 class Jobs extends Component {
@@ -54,18 +54,18 @@ class Jobs extends Component {
     }
       
 
-  filterBorough=(placeholder)=>{
-      let {dataJobs} = this.state
-        let obj = {
-            queens: dataJobs.filter(point => point.borough_community === 'Queens' && 'Long Island  City'),
-            manhattan: dataJobs.filter(point => point.borough_community === 'Manhattan' && 'New York'),
-            bronx: dataJobs.filter(point => point.borough_community === 'Bronx'),
-            si: dataJobs.filter(point => point.borough_community === 'Staten Island'),
-            brooklyn: dataJobs.filter(point => point.borough_community === 'Brooklyn'),
+//   filterBorough=(placeholder)=>{
+//       let {dataJobs} = this.state
+//         let obj = {
+//             queens: dataJobs.filter(point => point.borough_community === 'Queens' && 'Long Island  City'),
+//             manhattan: dataJobs.filter(point => point.borough_community === 'Manhattan' && 'New York'),
+//             bronx: dataJobs.filter(point => point.borough_community === 'Bronx'),
+//             si: dataJobs.filter(point => point.borough_community === 'Staten Island'),
+//             brooklyn: dataJobs.filter(point => point.borough_community === 'Brooklyn'),
             
-        }
-        return obj[placeholder]
-    }
+//         }
+//         return obj[placeholder]
+//     }
 
     componentDidMount=() => {
         this.dataJobs()
@@ -74,7 +74,7 @@ class Jobs extends Component {
 render() {  
     const {dataJobs } =this.state
     console.log('dataJobs' , dataJobs )
-    console.log('YO THIS BRONX ' ,this.filterBorough('bronx'))
+    // console.log('YO THIS BRONX ' ,this.filterBorough('bronx'))
     return(
         <div >
             <h1>Jobs</h1>
