+++ src/index.css
@@ -3,6 +3,8 @@
 @import url("https://fonts.googleapis.com/css?family=Lato");
 @import url('https://fonts.googleapis.com/css?family=Pacifico');
 
+@import url('https://fonts.googleapis.com/css?family=Noto+Serif+KR');
+
 [class*="fontawesome-"]:before {
   /* font-family: 'FontAwesome', sans-serif; */
   content: "\f09b";
@@ -13,13 +15,13 @@
 body {
   margin: 0;
   padding: 0;
-  font-family:'Droid Serif', serif;  
+  /* font-family:'Droid Serif', serif;   */
+  font-family: 'Noto Serif KR', sans-serif;
   display: grid;
   font-size: 16px;
-  /* background-color: #e0e0e0; */
-  background-image:url('natural_paper.png')
+  background-color: #F0F0F0;
 }
-.home{
+/* .home{
   margin-top:50px;
   margin-left:400px;
   border:0px solid red;
@@ -27,7 +29,7 @@ body {
   height:200px;
   align-content:center;
   -webkit-align-content: center; 
-}
+} */
 .map-container {
   height: 280px;
   width: 550px;
@@ -83,9 +85,9 @@ body {
 .HolyGrail-content {
   flex: 1;
 }
-a {
+/* a {
   color: white;
-}
+} */
 
 .HolyGrail-nav {
   order: -1;
